Move source location functions to common module
